refactor(reactivity): extract isRef to shared
