+++ packages/reactivity/src/ref.ts
@@ -8,10 +8,12 @@ import {
 import { DirtyLevels, TrackOpTypes, TriggerOpTypes } from './constants'
 import {
   type IfAny,
+  type Ref,
   hasChanged,
   isArray,
   isFunction,
   isObject,
+  isRef,
 } from '@vue/shared'
 import {
   isProxy,
@@ -27,18 +29,8 @@ import { ComputedRefImpl } from './computed'
 import { getDepFromReactive } from './reactiveEffect'
 import { warn } from './warning'
 
-declare const RefSymbol: unique symbol
+export type { Ref } from '@vue/shared'
 export declare const RawSymbol: unique symbol
-
-export interface Ref<T = any> {
-  value: T
-  /**
-   * Type differentiator only.
-   * We need this to be in public d.ts but don't want it to show up in IDE
-   * autocomplete, so we use a private Symbol instead.
-   */
-  [RefSymbol]: true
-}
 
 type RefBase<T> = {
   dep?: Dep
@@ -90,16 +82,7 @@ export function triggerRefValue(
   }
 }
 
-/**
- * Checks if a value is a ref object.
- *
- * @param r - The value to inspect.
- * @see {@link https://vuejs.org/api/reactivity-utilities.html#isref}
- */
-export function isRef<T>(r: Ref<T> | unknown): r is Ref<T>
-export function isRef(r: any): r is Ref {
-  return !!(r && r.__v_isRef === true)
-}
+export { isRef } from '@vue/shared'
 
 /**
  * Takes an inner value and returns a reactive and mutable ref object, which
